Nuke navigator/util
