+++ src/software/ai/navigator/BUILD
@@ -1,40 +1,16 @@
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "util",
-    srcs = ["util.cpp"],
-    hdrs = ["util.h"],
-    deps = [
-        "//software/ai/primitive:move_primitive",
-        "//software/geom",
-        "//software/ai/navigator/obstacle",
-        "//software/util/parameter:dynamic_parameters",
-        "//software/ai/intent:avoid_area",
-        "//software/world",
-        "@g3log",
-    ],
-)
-
-cc_test(
-    name = "util_test",
-    srcs = ["util_test.cpp"],
-    deps = [
-        ":util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_library(
     name = "navigator",
     srcs = ["navigator.cpp"],
     hdrs = ["navigator.h"],
     visibility = ["//visibility:public"],
     deps = [
-        ":util",
         "//software/ai/intent",
         "//software/ai/intent:all_intents",
         "//software/ai/intent:intent_visitor",
         "//software/ai/navigator/obstacle",
+        "//software/ai/navigator/obstacle:obstacle_generation",
         "//software/ai/navigator/path_manager:path_manager",
         "//software/ai/primitive:all_primitives",
         "//software/util/parameter:dynamic_parameters",
