+++ src/software/ai/navigator/navigator.h
@@ -6,8 +6,8 @@
 #include "software/ai/intent/intent.h"
 #include "software/ai/intent/intent_visitor.h"
 #include "software/ai/navigator/obstacle/obstacle.h"
+#include "software/ai/navigator/obstacle/obstacle_generation.h"
 #include "software/ai/navigator/path_manager/path_manager.h"
-#include "software/ai/navigator/util.h"
 #include "software/ai/primitive/all_primitives.h"
 #include "software/ai/primitive/primitive.h"
 #include "software/util/parameter/dynamic_parameters.h"
@@ -144,6 +144,22 @@ class Navigator : public IntentVisitor
      * @return A factor from 0 to 1 for how close p is to an enemy obstacle
      */
     double getEnemyObstacleProximityFactor(const Point &p);
+
+    /**
+     * Calculates the transition speed for the robot between two line segments
+     *
+     * Calculates the speed that the robot should be at when it is at the end of a
+     * given line segment in order to smoothly transition to another given line segment,
+     * given a final speed at the end of the two line segments
+     *
+     * @param p1, p2, p3 are 3 points that define two line segments that form a path
+     * @param final_speed is the intended final speed at the end of the path
+     * @return the first segment's final speed after travelling from p1 to p2
+     * for a smooth transition to the p2 to p3 path, scaled by the final speed at the end
+     * of the path
+     */
+    double calculateTransitionSpeedBetweenSegments(const Point &p1, const Point &p2,
+                                                   const Point &p3, double final_speed);
 
     /**
      * Registers this robot id as a robot that is not assigned a MoveIntent
