+++ src/software/ai/navigator/obstacle/obstacle_generation_test.cpp
@@ -1,0 +1,19 @@
+#include "software/ai/navigator/obstacle/obstacle_generation.h"
+
+#include <gtest/gtest.h>
+#include <math.h>
+
+#include <cmath>
+#include <cstdlib>
+#include <ctime>
+#include <sstream>
+
+#include "software/geom/point.h"
+#include "software/geom/util.h"
+
+int main(int argc, char **argv)
+{
+    std::cout << argv[0] << std::endl;
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
