--- src/software/ai/navigator/util.h
@@ -1,44 +1,0 @@
-#pragma once
-
-#include "software/ai/intent/avoid_area.h"
-#include "software/ai/navigator/obstacle/obstacle.h"
-#include "software/ai/primitive/move_primitive.h"
-#include "software/geom/point.h"
-#include "software/geom/util.h"
-#include "software/util/parameter/dynamic_parameters.h"
-#include "software/world/world.h"
-
-/**
- * Calculates the transition speed for the robot between two line segments
- *
- * Calculates the speed that the robot should be at when it is at the end of a
- * given line segment in order to smoothly transition to another given line segment,
- * given a final speed at the end of the two line segments
- *
- * @param p1, p2, p3 are 3 points that define two line segments that form a path
- * @param final_speed is the intended final speed at the end of the path
- * @return the first segment's final speed after travelling from p1 to p2
- * for a smooth transition to the p2 to p3 path, scaled by the final speed at the end
- * of the path
- */
-double calculateTransitionSpeedBetweenSegments(const Point &p1, const Point &p2,
-                                               const Point &p3, double final_speed);
-
-/**
- * Create obstacles for the given avoid areas, with a buffer such that the edge
- * of the robot does not protrude into the area
- *
- * @param avoid_areas The areas to convert into obstacles
- * @param world the world that's full of obstacles
- *
- * @return Obstacles representing the given avoid areas
- */
-std::vector<Obstacle> getObstaclesFromAvoidAreas(
-    const std::vector<AvoidArea> &avoid_areas, World world);
-
-/**
- * Get Obstacles from a Team
- *
- * @return vector of obstacles
- */
-std::vector<Obstacle> getObstaclesFromTeam(const Team &team);
